+++ src/framework/parsers/sogs.js
@@ -1,7 +1,8 @@
 import { Asset } from '../../framework/asset/asset.js';
 import { http, Http } from '../../platform/net/http.js';
-import { GSplatResource } from './gsplat-resource.js';
+import { GSplatResource } from '../../scene/gsplat/gsplat-resource.js';
 import { GSplatSogsData } from '../../scene/gsplat/gsplat-sogs-data.js';
+import { GSplatSogsResource } from '../../scene/gsplat/gsplat-sogs-resource.js';
 
 /**
  * @import { AppBase } from '../app-base.js'
@@ -74,7 +75,9 @@ class SogsParser {
             await data.reorderData();
         }
 
-        const resource = new GSplatResource(this.app, (asset.data?.decompress) ? (await data.decompress()) : data, []);
+        const resource = asset.data?.decompress ?
+            new GSplatResource(this.app .graphicsDevice, await data.decompress()) :
+                new GSplatSogsResource(this.app.graphicsDevice, data);
 
         callback(null, resource);
     }
